Moved BaseConfigView into common logistics module
